Source code cleanup
as a part of code enhancements
