+++ src/fheroes2/game/game_interface.cpp
@@ -30,7 +30,6 @@
 #include "game_interface.h"
 #include "maps.h"
 #include "mp2.h"
-#include "players.h"
 #include "settings.h"
 #include "ui_tool.h"
 #include "world.h"
